# opcua-webservice-node/src/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: office@opcfoundation.org
  description: Provides simple HTTPS based access to an OPC UA server.
  license:
    name: OPC Source Deliverable Agreement of Use
    url: https://opcfoundation.org/license/source/1.11/
  title: OPC UA Web API
  version: 1.05.4
servers:
- url: http://localhost:4840
paths:
...
  /translate:
    post:
      operationId: TranslateBrowsePathsToNodeIds
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranslateBrowsePathsToNodeIdsRequest"
        description: "[Link to specification](https://reference.opcfoundation.org/v105/Core/docs/Part4/5.9.4/#5.9.4.2)."
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslateBrowsePathsToNodeIdsResponse"
          description: "[Link to specification](https://reference.opcfoundation.org/v105/Core/docs/Part4/5.9.4/#5.9.4.2)."
      x-eov-operation-handler: controllers/DefaultController
...
  /findservers:
    post:
      operationId: FindServers
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindServersRequest"
        description: "[Link to specification](https://reference.opcfoundation.org/v105/Core/docs/Part4/5.5.2/#5.5.2.2)."
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindServersResponse"
          description: "[Link to specification](https://reference.opcfoundation.org/v105/Core/docs/Part4/5.5.2/#5.5.2.2)."
      x-eov-operation-handler: controllers/DefaultController
...
components:
  schemas:
    StatusCode:
      additionalProperties: false
      description: "[Link to specification](https://reference.opcfoundation.org/Core/Part4/v105/docs/7.39).\
        \ [Set of defined codes](https://github.com/OPCFoundation/UA-Nodeset/tree/latest/Schema/StatusCode.csv)."
      example:
        Symbol: Symbol
        Code: 2588772717
      properties:
        Code:
          format: int64
          maximum: 4294967295
          minimum: 0
          type: integer
        Symbol:
          type: string
      type: object
    LocalizedText:
      additionalProperties: false
      description: "[Link to specification](https://reference.opcfoundation.org/Core/Part3/v105/docs/8.5)."
      example:
        Locale: Locale
        Text: Text
      properties:
        Locale:
          format: rfc3066
          type: string
        Text:
          type: string
      type: object
    ExtensionObject:
      description: "[Link to specification](https://reference.opcfoundation.org/Core/Part6/v105/docs/5.4.2.16)."
      example:
        UaTypeId: UaTypeId
        UaEncoding: 180
        UaBody: UaBody
      properties:
        UaTypeId:
          format: UaNodeId
          type: string
        UaEncoding:
          format: int32
          maximum: 255
          minimum: 0
          type: integer
        UaBody:
          format: byte
          type: string
      type: object
...
    DiagnosticInfo:
      additionalProperties: false
      description: "[Link to specification](https://reference.opcfoundation.org/Core/Part4/v105/docs/7.12)."
      example:
        Locale: 5
        AdditionalInfo: AdditionalInfo
        NamespaceUri: 5
        SymbolicId: 1
        InnerDiagnosticInfo: null
        LocalizedText: 2
        InnerStatusCode:
          Symbol: Symbol
          Code: 2588772717
      properties:
        SymbolicId:
          format: int32
          type: integer
        NamespaceUri:
          format: int32
          type: integer
        Locale:
          format: int32
          type: integer
        LocalizedText:
          format: int32
          type: integer
        AdditionalInfo:
          type: string
        InnerStatusCode:
          $ref: "#/components/schemas/StatusCode"
        InnerDiagnosticInfo:
          $ref: "#/components/schemas/DiagnosticInfo"
      type: object
...
    ApplicationDescription:
      description: "[Link to specification](https://reference.opcfoundation.org/v105/Core/docs/Part14/7.2.4/#7.2.4.6.5)."
      example:
        ApplicationType: 0
        ApplicationUri: ApplicationUri
        ProductUri: ProductUri
        ApplicationName:
          Locale: Locale
          Text: Text
        DiscoveryUrls:
        - DiscoveryUrls
        - DiscoveryUrls
        DiscoveryProfileUri: DiscoveryProfileUri
        GatewayServerUri: GatewayServerUri
      properties:
        ApplicationUri:
          type: string
        ProductUri:
          type: string
        ApplicationName:
          $ref: "#/components/schemas/LocalizedText"
        ApplicationType:
          description: "[Link to specification](https://reference.opcfoundation.org/v105/Core/docs/Part4/7.4)."
          format: ApplicationType
          type: integer
        GatewayServerUri:
          type: string
        DiscoveryProfileUri:
          type: string
        DiscoveryUrls:
          items:
            type: string
          type: array
      type: object
    RequestHeader:
      description: "[Link to specification](https://reference.opcfoundation.org/v105/Core/docs/Part4/7.33)."
      example:
        AuthenticationToken: AuthenticationToken
        RequestHandle: 343953088
        AuditEntryId: AuditEntryId
        ReturnDiagnostics: 2588772717
        TimeoutHint: 629561881
        AdditionalHeader:
          UaTypeId: UaTypeId
          UaEncoding: 180
          UaBody: UaBody
        Timestamp: 2000-01-23T04:56:07.000+00:00
      properties:
        AuthenticationToken:
          format: UaNodeId
          type: string
        Timestamp:
          default: 0001-01-01T00:00:00Z
          format: date-time
          type: string
        RequestHandle:
          default: 0
          format: int64
          maximum: 4294967295
          minimum: 0
          type: integer
        ReturnDiagnostics:
          default: 0
          format: int64
          maximum: 4294967295
          minimum: 0
          type: integer
        AuditEntryId:
          type: string
        TimeoutHint:
          default: 0
          format: int64
          maximum: 4294967295
          minimum: 0
          type: integer
        AdditionalHeader:
          $ref: "#/components/schemas/ExtensionObject"
      type: object
    ResponseHeader:
      description: "[Link to specification](https://reference.opcfoundation.org/v105/Core/docs/Part4/7.34)."
      example:
        StringTable:
        - StringTable
        - StringTable
        ServiceResult:
          Symbol: Symbol
          Code: 2588772717
        RequestHandle: 343953088
        AdditionalHeader:
          UaTypeId: UaTypeId
          UaEncoding: 180
          UaBody: UaBody
        Timestamp: 2000-01-23T04:56:07.000+00:00
        ServiceDiagnostics:
          Locale: 5
          AdditionalInfo: AdditionalInfo
          NamespaceUri: 5
          SymbolicId: 1
          InnerDiagnosticInfo: null
          LocalizedText: 2
          InnerStatusCode:
            Symbol: Symbol
            Code: 2588772717
      properties:
        Timestamp:
          default: 0001-01-01T00:00:00Z
          format: date-time
          type: string
        RequestHandle:
          default: 0
          format: int64
          maximum: 4294967295
          minimum: 0
          type: integer
        ServiceResult:
          $ref: "#/components/schemas/StatusCode"
        ServiceDiagnostics:
          $ref: "#/components/schemas/DiagnosticInfo"
        StringTable:
          items:
            type: string
          type: array
        AdditionalHeader:
          $ref: "#/components/schemas/ExtensionObject"
      type: object
...
    FindServersRequest:
      description: "[Link to specification](https://reference.opcfoundation.org/v105/Core/docs/Part4/5.5.2/#5.5.2.2)."
      example:
        ServerUris:
        - ServerUris
        - ServerUris
        RequestHeader:
          AuthenticationToken: AuthenticationToken
          RequestHandle: 343953088
          AuditEntryId: AuditEntryId
          ReturnDiagnostics: 2588772717
          TimeoutHint: 629561881
          AdditionalHeader:
            UaTypeId: UaTypeId
            UaEncoding: 180
            UaBody: UaBody
          Timestamp: 2000-01-23T04:56:07.000+00:00
        EndpointUrl: EndpointUrl
        LocaleIds:
        - LocaleIds
        - LocaleIds
      properties:
        RequestHeader:
          $ref: "#/components/schemas/RequestHeader"
        EndpointUrl:
          type: string
        LocaleIds:
          items:
            type: string
          type: array
        ServerUris:
          items:
            type: string
          type: array
      type: object
    FindServersResponse:
      description: "[Link to specification](https://reference.opcfoundation.org/v105/Core/docs/Part4/5.5.2/#5.5.2.2)."
      example:
        ResponseHeader:
          StringTable:
          - StringTable
          - StringTable
          ServiceResult:
            Symbol: Symbol
            Code: 2588772717
          RequestHandle: 343953088
          AdditionalHeader:
            UaTypeId: UaTypeId
            UaEncoding: 180
            UaBody: UaBody
          Timestamp: 2000-01-23T04:56:07.000+00:00
          ServiceDiagnostics:
            Locale: 5
            AdditionalInfo: AdditionalInfo
            NamespaceUri: 5
            SymbolicId: 1
            InnerDiagnosticInfo: null
            LocalizedText: 2
            InnerStatusCode:
              Symbol: Symbol
              Code: 2588772717
        Servers:
        - ApplicationType: 0
          ApplicationUri: ApplicationUri
          ProductUri: ProductUri
          ApplicationName:
            Locale: Locale
            Text: Text
          DiscoveryUrls:
          - DiscoveryUrls
          - DiscoveryUrls
          DiscoveryProfileUri: DiscoveryProfileUri
          GatewayServerUri: GatewayServerUri
        - ApplicationType: 0
          ApplicationUri: ApplicationUri
          ProductUri: ProductUri
          ApplicationName:
            Locale: Locale
            Text: Text
          DiscoveryUrls:
          - DiscoveryUrls
          - DiscoveryUrls
          DiscoveryProfileUri: DiscoveryProfileUri
          GatewayServerUri: GatewayServerUri
      properties:
        ResponseHeader:
          $ref: "#/components/schemas/ResponseHeader"
        Servers:
          items:
            $ref: "#/components/schemas/ApplicationDescription"
          type: array
      type: object
...
    RelativePathElement:
      description: "[Link to specification](https://reference.opcfoundation.org/v105/Core/docs/Part4/7.31)."
      example:
        IncludeSubtypes: false
        IsInverse: false
        ReferenceTypeId: ReferenceTypeId
        TargetName: TargetName
      properties:
        ReferenceTypeId:
          format: UaNodeId
          type: string
        IsInverse:
          default: false
          type: boolean
        IncludeSubtypes:
          default: false
          type: boolean
        TargetName:
          format: UaQualifiedName
          type: string
      type: object
    RelativePath:
      description: "[Link to specification](https://reference.opcfoundation.org/v105/Core/docs/Part4/7.31)."
      example:
        Elements:
        - IncludeSubtypes: false
          IsInverse: false
          ReferenceTypeId: ReferenceTypeId
          TargetName: TargetName
        - IncludeSubtypes: false
          IsInverse: false
          ReferenceTypeId: ReferenceTypeId
          TargetName: TargetName
      properties:
        Elements:
          items:
            $ref: "#/components/schemas/RelativePathElement"
          type: array
      type: object
    BrowsePath:
      description: "[Link to specification](https://reference.opcfoundation.org/v105/Core/docs/Part4/5.9.4/#5.9.4.2)."
      example:
        StartingNode: StartingNode
        RelativePath:
          Elements:
          - IncludeSubtypes: false
            IsInverse: false
            ReferenceTypeId: ReferenceTypeId
            TargetName: TargetName
          - IncludeSubtypes: false
            IsInverse: false
            ReferenceTypeId: ReferenceTypeId
            TargetName: TargetName
      properties:
        StartingNode:
          format: UaNodeId
          type: string
        RelativePath:
          $ref: "#/components/schemas/RelativePath"
      type: object
    BrowsePathTarget:
      description: "[Link to specification](https://reference.opcfoundation.org/v105/Core/docs/Part4/5.9.4/#5.9.4.2)."
      example:
        TargetId: TargetId
        RemainingPathIndex: 343953088
      properties:
        TargetId:
          format: UaExpandedNodeId
          type: string
        RemainingPathIndex:
          default: 0
          format: int64
          maximum: 4294967295
          minimum: 0
          type: integer
      type: object
    BrowsePathResult:
      description: "[Link to specification](https://reference.opcfoundation.org/v105/Core/docs/Part4/5.9.4/#5.9.4.2)."
      example:
        Targets:
        - TargetId: TargetId
          RemainingPathIndex: 343953088
        - TargetId: TargetId
          RemainingPathIndex: 343953088
        StatusCode:
          Symbol: Symbol
          Code: 2588772717
      properties:
        StatusCode:
          $ref: "#/components/schemas/StatusCode"
        Targets:
          items:
            $ref: "#/components/schemas/BrowsePathTarget"
          type: array
      type: object
    TranslateBrowsePathsToNodeIdsRequest:
      description: "[Link to specification](https://reference.opcfoundation.org/v105/Core/docs/Part4/5.9.4/#5.9.4.2)."
      example:
        RequestHeader:
          AuthenticationToken: AuthenticationToken
          RequestHandle: 343953088
          AuditEntryId: AuditEntryId
          ReturnDiagnostics: 2588772717
          TimeoutHint: 629561881
          AdditionalHeader:
            UaTypeId: UaTypeId
            UaEncoding: 180
            UaBody: UaBody
          Timestamp: 2000-01-23T04:56:07.000+00:00
        BrowsePaths:
        - StartingNode: StartingNode
          RelativePath:
            Elements:
            - IncludeSubtypes: false
              IsInverse: false
              ReferenceTypeId: ReferenceTypeId
              TargetName: TargetName
            - IncludeSubtypes: false
              IsInverse: false
              ReferenceTypeId: ReferenceTypeId
              TargetName: TargetName
        - StartingNode: StartingNode
          RelativePath:
            Elements:
            - IncludeSubtypes: false
              IsInverse: false
              ReferenceTypeId: ReferenceTypeId
              TargetName: TargetName
            - IncludeSubtypes: false
              IsInverse: false
              ReferenceTypeId: ReferenceTypeId
              TargetName: TargetName
      properties:
        RequestHeader:
          $ref: "#/components/schemas/RequestHeader"
        BrowsePaths:
          items:
            $ref: "#/components/schemas/BrowsePath"
          type: array
      type: object
    TranslateBrowsePathsToNodeIdsResponse:
      description: "[Link to specification](https://reference.opcfoundation.org/v105/Core/docs/Part4/5.9.4/#5.9.4.2)."
      example:
        ResponseHeader:
          StringTable:
          - StringTable
          - StringTable
          ServiceResult:
            Symbol: Symbol
            Code: 2588772717
          RequestHandle: 343953088
          AdditionalHeader:
            UaTypeId: UaTypeId
            UaEncoding: 180
            UaBody: UaBody
          Timestamp: 2000-01-23T04:56:07.000+00:00
          ServiceDiagnostics:
            Locale: 5
            AdditionalInfo: AdditionalInfo
            NamespaceUri: 5
            SymbolicId: 1
            InnerDiagnosticInfo: null
            LocalizedText: 2
            InnerStatusCode:
              Symbol: Symbol
              Code: 2588772717
        Results:
        - Targets:
          - TargetId: TargetId
            RemainingPathIndex: 343953088
          - TargetId: TargetId
            RemainingPathIndex: 343953088
          StatusCode:
            Symbol: Symbol
            Code: 2588772717
        - Targets:
          - TargetId: TargetId
            RemainingPathIndex: 343953088
          - TargetId: TargetId
            RemainingPathIndex: 343953088
          StatusCode:
            Symbol: Symbol
            Code: 2588772717
        DiagnosticInfos:
        - Locale: 5
          AdditionalInfo: AdditionalInfo
          NamespaceUri: 5
          SymbolicId: 1
          InnerDiagnosticInfo: null
          LocalizedText: 2
          InnerStatusCode:
            Symbol: Symbol
            Code: 2588772717
        - Locale: 5
          AdditionalInfo: AdditionalInfo
          NamespaceUri: 5
          SymbolicId: 1
          InnerDiagnosticInfo: null
          LocalizedText: 2
          InnerStatusCode:
            Symbol: Symbol
            Code: 2588772717
      properties:
        ResponseHeader:
          $ref: "#/components/schemas/ResponseHeader"
        Results:
          items:
            $ref: "#/components/schemas/BrowsePathResult"
          type: array
        DiagnosticInfos:
          items:
            $ref: "#/components/schemas/DiagnosticInfo"
          type: array
      type: object
...